chore: Configures code style
With Editorconfig installed in VScode, with a right-click in the root of the project the plugin generates a .editorconfig file.

Installing eslint with

yarn add eslint -D

and running

yarn eslint --init

will configure eslint and generate a .eslintrc.js file.

Finnaly we just install prettier, create a .prettierrc file and configure eslint to use it.

yarn add prettier eslint-config-prettier eslint-plugin-prettier babel-eslint -D